requirements.txt
# MyMultiAgent Dependencies

# Core dependencies
pyyaml>=6.0

# Optional: for enhanced async support
# asyncio-mqtt>=0.16.0

# Optional: for better HTTP client
# httpx>=0.25.0

# Testing
pytest>=7.0.0

# Development
black>=23.0.0
ruff>=0.1.0
mypy>=1.0.0
